src/chainlit-demo/requirements.txt
langchain
langchain-community
chainlit
python-dotenv
openai
langchain-openai